src/main/java/org/junit/assessment/FileUtils.java
package org.junit.assessment;

import java.io.*;

public class FileUtils {

    public boolean writeData(String path,String word) {
        FileWriter fw = null;
        try {
            fw = new FileWriter(path);
            fw.write(word);
            fw.flush();
            return true;
        } catch (IOException ex) {
            ex.printStackTrace();
            return false;
        } finally {
            if (fw != null) {
                try {
                    fw.close();
                } catch (IOException ex) {
                    ex.printStackTrace();
                }
            }
        }
    }

    public String readData(String filePath) {
        StringBuilder content = new StringBuilder();
        try (BufferedReader reader = new BufferedReader(new FileReader(filePath))) {
            String line;
            while ((line = reader.readLine()) != null) {
                content.append(line).append("\n");
            }
        } catch (IOException e) {
            e.printStackTrace();
        }
        return content.toString();
    }
}
